requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9
pydantic>=2.7.0

# ── LINE Bot SDK ───────────────────────────────────────────────────────────────
line-bot-sdk>=3.11.0

# ── AI ────────────────────────────────────────────────────────────────────────
google-generativeai>=0.7.0

# ── HTTP / Storage ────────────────────────────────────────────────────────────
httpx>=0.27.0
Pillow>=10.3.0
pyyaml>=6.0.1
aiofiles>=23.2.1

# ── Config / Utilities ────────────────────────────────────────────────────────
python-dotenv>=1.0.1
nest_asyncio>=1.6.0
